app/requirements.txt
Flask==2.0.2
flask_login==0.6.2
Flask-SQLAlchemy==2.5.1
flask_bootstrap==3.3.7.1
datetime==4.5
wtforms==3.0.1
flask_wtf==1.0.1
email_validator==1.2.1
Flask-Migrate==3.1.0
python_decouple==3.6
requests==2.28.1
jwt==1.3.1
psycopg2==2.9.1